Specify a PipedriveEmailType and PipedrivePhonenumberType
Pipedrive is inconsistent with the values it returns for email addresses and phone numbers. It can return both string types and lists containing a dict containing strings. We've now made a PipedriveListDictStringOrStringType field we can use for these return values. Ugly? Yes. Effective? Yes. We inherent from this field so we have more logical names for the fields we want to use (being PipedriveEmailType and PipedrivePhonenumberType).